Player apply damage to Destroyable objects (#130)
* feat: api for character actions

* wip

* wip: colapsing flurry punches with player gameplay and Unity Events

* feat: api to interact with flurry punching animation events

* feat: set Ajax rigid body to never sleep && problem using unity events transformation from collider2d and object

* feat: simplifies interact 2d api && Ajax punches interact with Enemy's destroyable component

* feat: dash spear damage linked to player

+++ Erlang-Legacy/Assets/Scripts/Player/Stats.cs
@@ -13,9 +13,9 @@ namespace Core.Player
 
 
         [Header("Damage")]
-        public float dashDamage;
-        public float rayDamage; // how much damage ray does
-        public float punchDamage;
+        public int dashDamage;
+        public int rayDamage; // how much damage ray does
+        public int punchDamage;
 
 
         [Header("Cooldown or events duration")]
